refactor: move decode revert to evm
