+++ utils/Cargo.toml
@@ -19,7 +19,6 @@ eyre = { version = "0.6.5", default-features = false }
 hex = "0.4.3"
 reqwest = { version = "0.11.8", default-features = false, features = ["json", "rustls"] }
 rustc-hex = { version = "2.1.0", default-features = false }
-revm = {version = "1.5", default-features = false}
 serde = "1.0.132"
 serde_json = { version = "1.0.67", default-features = false }
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
